get all books using @RequestMapping
get book by @GetMapping

update book by id by @PutMapping with @RequestBody and @PathVaraiable

delete book by id by @DeleteMapping with @PathVaraiable

+++ src/main/java/com/example/demo/repository/BooksRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface BooksRepository extends JpaRepository<Book,Long> {
+public interface BooksRepository extends JpaRepository<Book,Integer> {
 
 
 }
